#crud sem validacoes

+++ src/test/java/com/hpv/CafeManha/CafeManhaApplicationTests.java
@@ -7,14 +7,14 @@ import java.util.Date;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
 
-import model.PratoColaboradorModel;
+import com.hpv.CafeManha.model.PratoColaboradorModel;
 
 @SpringBootTest
 class CafeManhaApplicationTests {
 
 	@Test
 	void contextLoads() {
-		PratoColaboradorModel pratoColaboradorModel=new PratoColaboradorModel(1,"Joao","99999999999","Cafe",new Date(),true);
+		PratoColaboradorModel pratoColaboradorModel=new PratoColaboradorModel((long) 1,"Joao","99999999999","Cafe",new Date(),true);
 		
 		assertEquals("Joao", pratoColaboradorModel.getNome());
 	}
